requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Dashboard and Visualization
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Data Storage (Optional)
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9  # pinned to avoid build issues on macOS arm64; optional (remove if not using Postgres)

# Web Scraping and API
selenium>=4.15.0
webdriver-manager>=4.0.0
aiohttp>=3.8.0

# Date and Time Processing
python-dateutil>=2.8.0
pytz>=2023.3

# Data Validation and Configuration
pydantic>=2.4.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Performance and Caching
redis>=5.0.0
joblib>=1.3.0

# Export and Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0
fpdf2>=2.7.0

# Logging and Monitoring
loguru>=0.7.0
